Ballooning event window and other changes
Ballooning decision to open the ballooning event window that can be used by the player (reduces number of decisions)

Decision to sponsor a balloonist will be available for AI so they don't have to open the event screen all the time.

Switsched around some stuff in the decisions to prevent a gazillion errors.