Create mendelejew_custom_elements_customizable_localization_experimental.txt
I was running into problems trying to create a bunch of custom elements, that being that each one had to have a distinct prefix, not only from the otl elements, but also each other. This file is an alternative version of the main custom elements localization file, and is intended to allow me to experiment with what I believe will be a solution to the issue, without putting the baseline custom element naming file at risk. If all goes well, this file will essentially be an upgraded version of that one by the time I'm done.